refactor Query Editor Form
