+++ app/client/src/pages/Editor/QueryEditor/EditorJSONtoForm.tsx
@@ -1,15 +1,10 @@
 import { useContext, useEffect, useState } from "react";
 import React, { useCallback } from "react";
 import type { InjectedFormProps } from "redux-form";
-import { Tag } from "@blueprintjs/core";
 import { noop } from "lodash";
 import type { Datasource } from "entities/Datasource";
 import { DatasourceStructureContext } from "entities/Datasource";
-import {
-  getPluginImages,
-  getPluginNameFromId,
-} from "@appsmith/selectors/entitiesSelector";
-import FormControl from "../FormControl";
+import { getPluginNameFromId } from "@appsmith/selectors/entitiesSelector";
 import {
   PluginName,
   type Action,
@@ -17,41 +12,16 @@ import {
   type SaaSAction,
 } from "entities/Action";
 import { useDispatch, useSelector } from "react-redux";
-import ActionNameEditor from "components/editorComponents/ActionNameEditor";
-import DropdownField from "components/editorComponents/form/fields/DropdownField";
-import type { ControlProps } from "components/formControls/BaseControl";
 import ActionSettings from "pages/Editor/ActionSettings";
-import log from "loglevel";
-import {
-  Button,
-  Icon,
-  Spinner,
-  Tab,
-  TabPanel,
-  Tabs,
-  TabsList,
-  Tooltip,
-} from "design-system";
+import { Button, Tab, TabPanel, Tabs, TabsList, Tooltip } from "design-system";
 import styled from "styled-components";
 import FormRow from "components/editorComponents/FormRow";
 import { ResizerCSS } from "components/editorComponents/Debugger/Resizer";
 import AnalyticsUtil from "utils/AnalyticsUtil";
 import {
-  checkIfSectionCanRender,
-  checkIfSectionIsEnabled,
-  extractConditionalOutput,
-  isHidden,
-  modifySectionConfig,
-  updateEvaluatedSectionConfig,
-} from "components/formControls/utils";
-import {
-  ACTION_EDITOR_REFRESH,
-  CREATE_NEW_DATASOURCE,
   createMessage,
   DOCUMENTATION,
   DOCUMENTATION_TOOLTIP,
-  INVALID_FORM_CONFIGURATION,
-  UNEXPECTED_ERROR,
 } from "@appsmith/constants/messages";
 import { useParams } from "react-router";
 import type { AppState } from "@appsmith/reducers";
@@ -61,16 +31,13 @@ import ActionRightPane, {
 } from "components/editorComponents/ActionRightPane";
 import type { ActionResponse } from "api/ActionAPI";
 import type { Plugin } from "api/PluginApi";
-import { UIComponentTypes } from "api/PluginApi";
-import * as Sentry from "@sentry/react";
+import type { UIComponentTypes } from "api/PluginApi";
 import { DEBUGGER_TAB_KEYS } from "components/editorComponents/Debugger/helpers";
-import { EDITOR_TABS, SQL_DATASOURCES } from "constants/QueryEditorConstants";
+import { EDITOR_TABS } from "constants/QueryEditorConstants";
 import type { FormEvalOutput } from "reducers/evaluationReducers/formEvaluationReducer";
-import { isValidFormConfig } from "reducers/evaluationReducers/formEvaluationReducer";
 import { getQueryPaneConfigSelectedTabIndex } from "selectors/queryPaneSelectors";
 import { setQueryPaneConfigSelectedTabIndex } from "actions/queryPaneActions";
 import { ActionExecutionResizerHeight } from "pages/Editor/APIEditor/constants";
-import { getCurrentAppWorkspace } from "@appsmith/selectors/workspaceSelectors";
 import {
   getDebuggerSelectedTab,
   showDebuggerFlag,
@@ -78,19 +45,16 @@ import {
 import type { SourceEntity } from "entities/AppsmithConsole";
 import { ENTITY_TYPE as SOURCE_ENTITY_TYPE } from "entities/AppsmithConsole";
 import { DocsLink, openDoc } from "../../../constants/DocumentationLinks";
-import {
-  getHasCreateDatasourcePermission,
-  getHasExecuteActionPermission,
-  getHasManageActionPermission,
-} from "@appsmith/utils/BusinessFeatures/permissionPageHelpers";
+import { getHasExecuteActionPermission } from "@appsmith/utils/BusinessFeatures/permissionPageHelpers";
 import { useFeatureFlag } from "utils/hooks/useFeatureFlag";
 import { FEATURE_FLAG } from "@appsmith/entities/FeatureFlag";
 import { QueryEditorContext } from "./QueryEditorContext";
 import QueryResponseTabView from "./QueryResponseView";
 import { setDebuggerSelectedTab, showDebugger } from "actions/debuggerActions";
 import useShowSchema from "components/editorComponents/ActionRightPane/useShowSchema";
-import { isAppsmithAIPlugin } from "utils/editorContextUtils";
 import { doesPluginRequireDatasource } from "@appsmith/entities/Engine/actionHelpers";
+import FormRender from "./FormRender";
+import QueryEditorHeader from "./QueryEditorHeader";
 
 const QueryFormContainer = styled.form`
   flex: 1;
@@ -115,13 +79,6 @@ const QueryFormContainer = styled.form`
   }
 `;
 
-const ErrorMessage = styled.p`
-  font-size: 14px;
-  color: var(--ads-v2-color-fg-error);
-  display: inline-block;
-  margin-right: 10px;
-`;
-
 export const TabbedViewContainer = styled.div`
   ${ResizerCSS};
   height: ${ActionExecutionResizerHeight}px;
@@ -137,10 +94,6 @@ const SettingsWrapper = styled.div`
   height: 100%;
 `;
 
-const FieldWrapper = styled.div`
-  margin-top: 15px;
-`;
-
 const SecondaryWrapper = styled.div`
   display: flex;
   flex-direction: column;
@@ -152,42 +105,6 @@ export const StyledFormRow = styled(FormRow)`
   padding: 0px var(--ads-v2-spaces-7) var(--ads-v2-spaces-5)
     var(--ads-v2-spaces-7);
   flex: 0;
-`;
-
-const NameWrapper = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  input {
-    margin: 0;
-    box-sizing: border-box;
-  }
-`;
-
-const ActionsWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  flex: 1 1 50%;
-  justify-content: flex-end;
-  gap: var(--ads-v2-spaces-3);
-`;
-
-const DropdownSelect = styled.div`
-  font-size: 14px;
-  width: 230px;
-`;
-
-const CreateDatasource = styled.div`
-  display: flex;
-  gap: 8px;
-`;
-
-const StyledSpinner = styled(Spinner)`
-  display: flex;
-  height: 2vw;
-  align-items: center;
-  justify-content: space-between;
-  width: 5vw;
 `;
 
 const TabContainerView = styled.div`
@@ -316,9 +233,7 @@ export function EditorJSONtoForm(props: Props) {
     actionRightPaneAdditionSections,
     actionRightPaneBackLink,
     closeEditorLink,
-    moreActionsMenu,
     notification,
-    saveActionName,
     showSuggestedWidgets = true,
   } = useContext(QueryEditorContext);
 
@@ -334,22 +249,9 @@ export function EditorJSONtoForm(props: Props) {
   const [showResponseOnFirstLoad, setShowResponseOnFirstLoad] =
     useState<boolean>(false);
 
-  const isChangePermitted = getHasManageActionPermission(
-    isFeatureEnabled,
-    currentActionConfig?.userPermissions,
-  );
   const isExecutePermitted = getHasExecuteActionPermission(
     isFeatureEnabled,
     currentActionConfig?.userPermissions,
-  );
-
-  const userWorkspacePermissions = useSelector(
-    (state: AppState) => getCurrentAppWorkspace(state).userPermissions ?? [],
-  );
-
-  const canCreateDatasource = getHasCreateDatasourcePermission(
-    isFeatureEnabled,
-    userWorkspacePermissions,
   );
 
   const pluginRequireDatasource = doesPluginRequireDatasource(plugin);
@@ -367,23 +269,6 @@ export function EditorJSONtoForm(props: Props) {
   const currentActionPluginName = useSelector((state: AppState) =>
     getPluginNameFromId(state, currentActionConfig?.pluginId || ""),
   );
-
-  let actionBody = "";
-  if (!!currentActionConfig?.actionConfiguration) {
-    if ("formData" in currentActionConfig?.actionConfiguration) {
-      // if the action has a formData (the action is postUQI e.g. Oracle)
-      actionBody =
-        currentActionConfig.actionConfiguration.formData?.body?.data || "";
-    } else {
-      // if the action is pre UQI, the path is different e.g. mySQL
-      actionBody = currentActionConfig.actionConfiguration?.body || "";
-    }
-  }
-
-  // if (the body is empty and the action is an sql datasource) or the user does not have permission, block action execution.
-  const blockExecution =
-    (!actionBody && SQL_DATASOURCES.includes(currentActionPluginName)) ||
-    !isExecutePermitted;
 
   const dispatch = useDispatch();
 
@@ -417,148 +302,6 @@ export function EditorJSONtoForm(props: Props) {
     openDoc(DocsLink.QUERY, plugin?.documentationLink, plugin?.name);
   };
 
-  // Added function to handle the render of the configs
-  const renderConfig = (editorConfig: any) => {
-    try {
-      // Selectively rendering form based on uiComponent prop
-      if (uiComponent === UIComponentTypes.UQIDbEditorForm) {
-        // If the formEvaluation is not ready yet, just show loading state.
-        if (
-          props.hasOwnProperty("formEvaluationState") &&
-          !!props.formEvaluationState &&
-          Object.keys(props.formEvaluationState).length > 0
-        ) {
-          return editorConfig.map((config: any, idx: number) => {
-            return renderEachConfigV2(formName, config, idx);
-          });
-        } else {
-          return <StyledSpinner size="md" />;
-        }
-      } else {
-        return editorConfig.map(renderEachConfig(formName));
-      }
-    } catch (e) {
-      log.error(e);
-      Sentry.captureException(e);
-      return (
-        <>
-          <ErrorMessage>
-            {createMessage(INVALID_FORM_CONFIGURATION)}
-          </ErrorMessage>
-          <Tag
-            intent="warning"
-            interactive
-            minimal
-            onClick={() => window.location.reload()}
-            round
-          >
-            {createMessage(ACTION_EDITOR_REFRESH)}
-          </Tag>
-        </>
-      );
-    }
-  };
-
-  // Render function to render the V2 of form editor type (UQI)
-  // Section argument is a nested config object, this function recursively renders the UI based on the config
-  const renderEachConfigV2 = (formName: string, section: any, idx: number) => {
-    let enabled = true;
-    if (!!section) {
-      // If the section is a nested component, recursively check for conditional statements
-      if (
-        "schema" in section &&
-        Array.isArray(section.schema) &&
-        section.schema.length > 0
-      ) {
-        section.schema = section.schema.map((subSection: any) => {
-          const conditionalOutput = extractConditionalOutput(
-            subSection,
-            props.formEvaluationState,
-          );
-          if (!checkIfSectionCanRender(conditionalOutput)) {
-            subSection.hidden = true;
-          } else {
-            subSection.hidden = false;
-          }
-          enabled = checkIfSectionIsEnabled(conditionalOutput);
-          subSection = updateEvaluatedSectionConfig(
-            subSection,
-            conditionalOutput,
-            enabled,
-          );
-          if (!isValidFormConfig(subSection)) return null;
-          return subSection;
-        });
-      }
-      // If the component is not allowed to render, return null
-      const conditionalOutput = extractConditionalOutput(
-        section,
-        props.formEvaluationState,
-      );
-      if (!checkIfSectionCanRender(conditionalOutput)) return null;
-      section = updateEvaluatedSectionConfig(section, conditionalOutput);
-      enabled = checkIfSectionIsEnabled(conditionalOutput);
-      if (!isValidFormConfig(section)) return null;
-    }
-    if (section.hasOwnProperty("controlType")) {
-      // If component is type section, render it's children
-      if (
-        section.controlType === "SECTION" &&
-        section.hasOwnProperty("children")
-      ) {
-        return section.children.map((section: any, idx: number) => {
-          return renderEachConfigV2(formName, section, idx);
-        });
-      }
-      try {
-        const { configProperty } = section;
-        const modifiedSection = modifySectionConfig(section, enabled);
-        return (
-          <FieldWrapper key={`${configProperty}_${idx}`}>
-            <FormControl config={modifiedSection} formName={formName} />
-          </FieldWrapper>
-        );
-      } catch (e) {
-        log.error(e);
-      }
-    } else {
-      return section.map((section: any, idx: number) => {
-        renderEachConfigV2(formName, section, idx);
-      });
-    }
-    return null;
-  };
-
-  // Recursive call to render forms pre UQI
-  const renderEachConfig =
-    (formName: string) =>
-    (section: any): any => {
-      return section.children.map(
-        (formControlOrSection: ControlProps, idx: number) => {
-          if (isHidden(props.formData, section.hidden, undefined, false))
-            return null;
-          if (formControlOrSection.hasOwnProperty("children")) {
-            return renderEachConfig(formName)(formControlOrSection);
-          } else {
-            try {
-              const { configProperty } = formControlOrSection;
-              return (
-                <FieldWrapper key={`${configProperty}_${idx}`}>
-                  <FormControl
-                    config={formControlOrSection}
-                    formName={formName}
-                  />
-                </FieldWrapper>
-              );
-            } catch (e) {
-              log.error(e);
-            }
-          }
-          return null;
-        },
-      );
-    };
-
   const responseTabOnRunClick = () => {
     props.onRunClick();
 
@@ -578,30 +321,6 @@ export function EditorJSONtoForm(props: Props) {
 
   const { hasDependencies } = useEntityDependencies(props.actionName);
 
-  const pluginImages = useSelector(getPluginImages);
-
-  interface DATASOURCES_OPTIONS_TYPE {
-    label: string;
-    value: string;
-    image: string;
-  }
-
-  // Filtering the datasources for listing the similar datasources only rather than having all the active datasources in the list, which on switching resulted in error.
-  const DATASOURCES_OPTIONS: Array<DATASOURCES_OPTIONS_TYPE> =
-    dataSources.reduce(
-      (acc: Array<DATASOURCES_OPTIONS_TYPE>, dataSource: Datasource) => {
-        if (dataSource.pluginId === plugin?.id) {
-          acc.push({
-            label: dataSource.name,
-            value: dataSource.id,
-            image: pluginImages[dataSource.pluginId],
-          });
-        }
-        return acc;
-      },
-      [],
-    );
-
   const selectedConfigTab = useSelector(getQueryPaneConfigSelectedTabIndex);
 
   // Debugger render flag
@@ -620,11 +339,6 @@ export function EditorJSONtoForm(props: Props) {
     !!actionResponse ||
     currentActionPluginName !== PluginName.SMTP;
 
-  // Datasource selection is hidden for Appsmith AI Plugin and for plugins that don't require datasource
-  // TODO: @Diljit Remove this condition when knowledge retrieval for Appsmith AI is implemented (Only remove the AI Condition)
-  const showDatasourceSelector =
-    !isAppsmithAIPlugin(plugin?.packageName) && pluginRequireDatasource;
-
   // when switching between different redux forms, make sure this redux form has been initialized before rendering anything.
   // the initialized prop below comes from redux-form.
   if (!props.initialized) {
@@ -635,51 +349,14 @@ export function EditorJSONtoForm(props: Props) {
     <>
       {closeEditorLink}
       <QueryFormContainer onSubmit={handleSubmit(noop)}>
-        <StyledFormRow>
-          <NameWrapper>
-            <ActionNameEditor
-              disabled={!isChangePermitted}
-              saveActionName={saveActionName}
-            />
-          </NameWrapper>
-          <ActionsWrapper>
-            {moreActionsMenu}
-            {showDatasourceSelector && (
-              <DropdownSelect>
-                <DropdownField
-                  className={"t--switch-datasource"}
-                  formName={formName}
-                  isDisabled={!isChangePermitted}
-                  name="datasource.id"
-                  options={DATASOURCES_OPTIONS}
-                  placeholder="Datasource"
-                >
-                  {canCreateDatasource && (
-                    // this additional div is here so that rc-select can render the child with the onClick correctly
-                    <div>
-                      <CreateDatasource
-                        onClick={() => onCreateDatasourceClick()}
-                      >
-                        <Icon className="createIcon" name="plus" size="md" />
-                        {createMessage(CREATE_NEW_DATASOURCE)}
-                      </CreateDatasource>
-                    </div>
-                  )}
-                </DropdownField>
-              </DropdownSelect>
-            )}
-            <Button
-              className="t--run-query"
-              data-guided-tour-iid="run-query"
-              isDisabled={blockExecution}
-              isLoading={isRunning}
-              onClick={onRunClick}
-              size="md"
-            >
-              Run
-            </Button>
-          </ActionsWrapper>
-        </StyledFormRow>
+        <QueryEditorHeader
+          dataSources={dataSources}
+          formName={formName}
+          isRunning={isRunning}
+          onCreateDatasourceClick={onCreateDatasourceClick}
+          onRunClick={onRunClick}
+          plugin={plugin}
+        />
         {notification}
         <Wrapper>
           <div className="flex flex-1">
@@ -710,24 +387,13 @@ export function EditorJSONtoForm(props: Props) {
                     value={EDITOR_TABS.QUERY}
                   >
                     <SettingsWrapper>
-                      {editorConfig && editorConfig.length > 0 ? (
-                        renderConfig(editorConfig)
-                      ) : (
-                        <>
-                          <ErrorMessage>
-                            {createMessage(UNEXPECTED_ERROR)}
-                          </ErrorMessage>
-                          <Tag
-                            intent="warning"
-                            interactive
-                            minimal
-                            onClick={() => window.location.reload()}
-                            round
-                          >
-                            {createMessage(ACTION_EDITOR_REFRESH)}
-                          </Tag>
-                        </>
-                      )}
+                      <FormRender
+                        editorConfig={editorConfig}
+                        formData={props.formData}
+                        formEvaluationState={props.formEvaluationState}
+                        formName={formName}
+                        uiComponent={uiComponent}
+                      />
                     </SettingsWrapper>
                   </TabPanelWrapper>
                   <TabPanelWrapper value={EDITOR_TABS.SETTINGS}>
